# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "manhtb1982@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "Lấy danh sách categories",
                "description": "Lấy danh sách categories",
                "operationId": "indexCategory",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Phân trang",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "1"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Tên danh mục",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "categories"
                ],
                "summary": "Tạo category",
                "description": "Tạo category",
                "operationId": "storeCategory",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Category name",
                                        "type": "string",
                                        "example": "Mạnh"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "status": "success",
                                        "data": {
                                            "name": "Mạnh",
                                            "updated_at": "2022-11-21T04:25:14.000000Z",
                                            "created_at": "2022-11-21T04:25:14.000000Z",
                                            "id": 37
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "Lấy chi tiết category",
                "description": "Lấy chi tiết category",
                "operationId": "showCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id find",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "4"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "status": "success",
                                        "data": {
                                            "id": 4,
                                            "name": "Áo Nữ",
                                            "updated_at": "2022-11-21T04:25:14.000000Z",
                                            "created_at": "2022-11-21T04:25:14.000000Z"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "categories"
                ],
                "summary": "Update category",
                "description": "Update category",
                "operationId": "updateCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id category",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "1"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "User name",
                                        "type": "string",
                                        "example": "Mạnh"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "status": "success",
                                        "user": {
                                            "id": 30,
                                            "name": "Mạnh",
                                            "updated_at": "2022-11-21T04:25:14.000000Z",
                                            "created_at": "2022-11-21T04:25:14.000000Z"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Bad Request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "categories"
                ],
                "summary": "Xóa category",
                "description": "Xóa category",
                "operationId": "destroyCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id find",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "38"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "status": "success"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "categories"
                ],
                "summary": "Update category",
                "description": "Update category",
                "operationId": "updatePatchCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id category",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "30"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "User name",
                                        "type": "string",
                                        "example": "Mạnh"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "status": "success",
                                        "data": {
                                            "id": 30,
                                            "name": "Mạnh",
                                            "updated_at": "2022-11-21T04:25:14.000000Z",
                                            "created_at": "2022-11-21T04:25:14.000000Z"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}